feat: Update GainCoupon component to show coupon after a certain time
The GainCoupon component is modified to display a coupon after a specific amount of time has passed. This is achieved by adding a new state variable, 'showCoupon', and setting it to true after a delay of 15 seconds. The coupon is then rendered conditionally based on the value of 'showCoupon'.

Also, update the styling of the finish modal in the GainCoupon component to accommodate the new coupon display.

Fixes #17